chore: remove max width for HQ, extract alignment class to component and fix naming
