+++ src/components/max-desktop-width-item.vue
@@ -1,4 +1,4 @@
-<template functional>
+<template>
   <div
     class="x-layout-max-width-md x-layout-item desktop:x-layout-min-margin-32 large:x-layout-max-width-lg large:x-layout-min-margin-48"
   >
